refactor: implement feature-based architecture
- migrated auth components to (auth)/_components and _lib

- migrated onboarding components to onboarding/_components and _lib

- added docs/guide to folder structure
