+++ app/(app)/home/_components/achievements-section.tsx
