+++ app/(app)/home/_components/activity-card.tsx
