+++ app/(app)/home/_components/assistant-hero.tsx
