+++ app/(app)/home/_components/explore-section.tsx
