+++ app/(app)/home/_components/recent-activities.tsx
