+++ app/onboarding/_components/assistant-selection-form.tsx
