+++ app/onboarding/_components/guided-intro-completion.tsx
